feat: api 코드 이동 및 Dropdown 스토리북 경로 수정, 허스키 pre-commit 수정
