+++ packages/utils/src/apis/duri/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './home';
 export * from './auth';
-export * from './pet'
 export * from './quotation'
 export * from './test';
+export * from './request'
