--- packages/utils/src/apis/duri/hooks/pet.ts
@@ -1,12 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-
-import { getPetInfo } from '../pet';
-
-export const useGetPetInfo = () => {
-  const { data } = useQuery({
-    queryKey: ['getPetInfo'],
-    queryFn: () => getPetInfo(),
-    staleTime: 1000 * 60 * 10,
-  });
-  return data;
-};
