+++ packages/utils/src/apis/duri/index.ts
@@ -3,5 +3,4 @@ export * from './request';
 export * from './test';
 export * from './auth';
 export * from './hooks';
-export * from './pet';
 export * from './home';
