+++ packages/utils/src/apis/duri/request.ts
@@ -2,6 +2,15 @@
 import { duriInstance } from '../axiosConfig';
 import { RequestProps } from '../types/quotation';
 
+export async function getPetInfo() {
+  try {
+    const response = await duriInstance.get(`pet`);
+    return response.data;
+  } catch (error) {
+    console.error(error);
+  }
+}
+
 export const postRequestQuotation = async (request: RequestProps) => {
   // ): Promise<RequestResponse> => {
   const response = await duriInstance.post(`quotation/request`, request);
